新功能（现在有终端专用的低配版了）
