--- jianrongmode.cpp
@@ -1,1 +1,0 @@
-#i
