+++ namespace-master-gui.cpp
@@ -1,0 +1,31 @@
+#include <graphics.h>
+namespace master {
+	//const wchar_t* ziti = L"Arial"; //使用字体
+	char ziti[] = "Arial"; //使用字体
+	int a = GetSystemMetrics(SM_CXSCREEN); //这个变量int a是用户显示屏高
+	int b = a /4; //这个变量int b是程序运行时的高和宽
+	// 检查鼠标是否在按钮区域内
+    bool c (int x, int y, int btnX, int btnY, int btnWidth, int btnHeight) {
+    return x >= btnX && x <= btnX + btnWidth && y >= btnY && y <= btnY + btnHeight;
+	}
+	bool sa (int x, int y, int btnX1, int btnY, int btnWidth, int btnHeight) {
+		return x >= btnX1 && x <= btnX1 + btnWidth && y >= btnY && y <= btnY + btnHeight;
+		}
+	//如果鼠标在按钮区域内，返回布尔值c true，否则返回false
+	int qidongyouxianniudeX = b / 3 - 50; //a和b按钮的X坐标
+	int qidongyouxianniudeY = b / 3 * 2; //a和b按钮的Y坐标
+	void clearButton(int x, int y, int width, int height) {
+		setfillcolor(WHITE); // 设置填充颜色为白色
+		solidrectangle(x, y, x + width, y + height); // 用白色填充按钮区域
+	}
+	void chongzhipingmu() {
+		setfillcolor(WHITE); // 设置填充颜色为白色
+		solidrectangle(0, 0, b, b); // 用白色填充整个屏幕
+		IMAGE img; //定义一个图片对象
+		const char * a91 = "ATC4.ico"; //图片路径
+		loadimage(&img , a91 , b , b ,false); //加载图片
+		putimage(0,0 , &img ); //在屏幕上显示图片
+	}
+	int qidongqitajichangdeX = b / 3 * 2 - 50; //a和b按钮的X坐标
+	int qidongqitajichangdeY = b / 3 * 2; //a和b按钮的Y坐标
+}
